+++ resources/views/dashboard/events/trash.blade.php
@@ -1,11 +1,11 @@
 @extends('layouts.dashboard.master')
 
 @section('page-title')
-    Trash events ({{ \App\Models\Event::onlyTrashed()->count() }})
+    Trashed events ({{ \App\Models\Event::onlyTrashed()->count() }})
 @endsection
 
 @section('title')
-    Events ({{ \App\Models\Event::onlyTrashed()->count() }})
+    Trashed events ({{ \App\Models\Event::onlyTrashed()->count() }})
 @endsection
 
 @section('page-title-1', 'Events')
@@ -40,11 +40,13 @@
                                 <th>#</th>
                                 <th>ID</th>
                                 <th>Name</th>
-                                <th>Start</th>
-                                <th>End</th>
+                                <th>Organizer</th>
+                                <th>Status</th>
                                 <th>Venue</th>
-                                <td>Deleted</td>
-                                <th>Actions</th>
+                                @if (Auth::user()->role === 'admin')
+                                    <th>Actions</th>
+                                @endif
+                                <th>Deleted at</th>
                             </tr>
                         </thead>
                         <tbody>
@@ -52,39 +54,42 @@
                                 <tr>
                                     <td>{{ $loop->iteration }}</td>
                                     <td>{{ $event->id }}</td>
-                                    <td>{{ Str::words($event->name, 2, '...') }}</td>
-                                    <td>{{ \Carbon\Carbon::createFromFormat('Y-m-d H:i:s', $event->start_time)->format('y M, d H:i') }}</td>
-                                    <td>{{ \Carbon\Carbon::createFromFormat('Y-m-d H:i:s', $event->end_time)->format('y M, d H:i') }}</td>
+                                    <td>{{ $event->organizer->fullName() }}</td>
+                                    <td>{{ $event->status }}</td>
                                     <td>
-                                        This is a venue
+                                        {{ $event->venue->name }}
                                     </td>
                                     <td>
                                         {{ $event->deleted_at->diffForHumans() }}
                                     </td>
+                                    @if (Auth::user()->role === 'admin')
+                                        <td>
+                                            <form
+                                                method="POST"
+                                                action="{{ route('dashboard.events.restore', $event->id ) }}"
+                                                class="my-1"
+                                                style="display: inline-block"
+                                                > @csrf
+                                                    <button type="submit" class="btn btn-sm btn-success" style="font-size: 0.75rem;">Restore</button>
+                                            </form>
+                                            <form
+                                                method="POST"
+                                                action="{{ route('dashboard.events.forceDelete', $event->id ) }}"
+                                                class="my-1"
+                                                style="display: inline-block"
+                                                > @csrf
+                                                @method('DELETE')
+                                                    <button type="submit" class="btn btn-sm btn-danger" style="font-size: 0.75rem;" onclick="return confirm('Are you sure that you want to delete ({{ $event->name }}) permanently?');">Delete permanently</button>
+                                            </form>
+                                        </td>
+                                    @endif
                                     <td>
-                                        {{-- <a href="{{ route('dashboard.events.restore', [ 'event' => $event->slug]) }}" class="btn btn-sm btn-success" style="font-size: 0.75rem;">restore</a> --}}
-                                        <form
-                                            method="POST"
-                                            action="{{ route('dashboard.events.restore', $event->id ) }}"
-                                            class="my-1"
-                                            style="display: inline-block"
-                                            > @csrf
-                                                <button type="submit" class="btn btn-sm btn-success" style="font-size: 0.75rem;">Restore</button>
-                                        </form>
-                                        <form
-                                            method="POST"
-                                            action="{{ route('dashboard.events.forceDelete', $event->id ) }}"
-                                            class="my-1"
-                                            style="display: inline-block"
-                                            > @csrf
-                                            @method('DELETE')
-                                                <button type="submit" class="btn btn-sm btn-danger" style="font-size: 0.75rem;" onclick="return confirm('Are you sure that you want to delete ({{ $event->name }}) permanently?');">Delete permanently</button>
-                                        </form>
+                                        {{ $event->deleted_at }}
                                     </td>
                                 </tr>
                             @empty
                                 <tr>
-                                    <td colspan="5">
+                                    <td colspan="8">
                                         <div class="alert alert-info text-center">
                                             No events found.
                                         </div>
